requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-redis-cache==2.1.1
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.4.4
pytz==2019.3
redis==3.4.1
sentry-sdk==0.14.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.11.2
drf-yasg==1.17.1
